chore(gh workflow): concurrency key is now per branch instead of actors (#112)
With the per actor concurrency check, the same actor triggering CI in
two different PRs would have one cancelled in favor of the other. A side
of effect of that, is the cancellation does not prevent integration of
branch, the tests are just skipped.

Using the reference documentation of github for concurrency, this PR
uses the github.ref as the key. The `cancel-in-progress: true` ensures
that only the latest version of the commit in the CI is the choosen one.